+++ packages/eyes-sdk-core/test/unit/server/ServerConnector.spec.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const assert = require('assert')
-const fakeEyesServer = require('@applitools/sdk-fake-eyes-server')
+const {startFakeEyesServer} = require('@applitools/sdk-fake-eyes-server')
 const {ServerConnector, Logger, Configuration, GeneralUtils} = require('../../../')
 const {presult} = require('../../../lib/troubleshoot/utils')
 const logger = new Logger(process.env.APPLITOOLS_SHOW_LOGS)
@@ -16,7 +16,7 @@ function getServerConnector(config = {}) {
 
 describe('ServerConnector', () => {
   it('sends startSession request', async () => {
-    const {port, close} = await fakeEyesServer({logger})
+    const {port, close} = await startFakeEyesServer({logger})
     try {
       const serverUrl = `http://localhost:${port}`
       const serverConnector = getServerConnector({serverUrl})
@@ -48,7 +48,7 @@ describe('ServerConnector', () => {
 
   // [trello] https://trello.com/c/qjmAw1Sc/160-storybook-receiving-an-inconsistent-typeerror
   it("doesn't throw exception on server failure", async () => {
-    const {port, close} = await fakeEyesServer({logger, hangUp: true})
+    const {port, close} = await startFakeEyesServer({logger, hangUp: true})
     try {
       const serverUrl = `http://localhost:${port}`
       const serverConnector = getServerConnector({serverUrl})
@@ -60,7 +60,7 @@ describe('ServerConnector', () => {
   })
 
   it('getUserAgents works', async () => {
-    const {port, close} = await fakeEyesServer({logger})
+    const {port, close} = await startFakeEyesServer({logger})
     try {
       const serverUrl = `http://localhost:${port}`
       const serverConnector = getServerConnector({serverUrl})